Completely and majorly overhauled Breakin' Feds
Expect a massive improvement to framerate as this is a really, really thorough setup compared to normal. I personally now get 144 FPS consistently across the whole map

This is actually the third optimized version I've done for this map, number one was the version in the mod before and was the first map i gave an optimization pass to. It wasn't intended to even be publicly released and didn't fit my current standards. Version two was an improved version one though i lost the map files for it, still 1-2 months prior to this mod being released and now it has been overhauled to my current standards. 3 portals to 20 with a major performance gain to match.